backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend api"
readme = "README.md"
authors = [
    { name = "robert", email = "freshskates@proton.me" }
]
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.11",
    "prisma>=0.15.0",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
]

# uv run task <task-name>
[tool.taskipy.tasks]
build = "docker build -t backend-api ."
dev = "fastapi dev ./src/backend"
start = "fastapi run --workers 2 ./src/backend"
compose = "docker-compose up"
composebuild = "docker-compose up --build"
clean = "docker rmi backend-api"
test = "pytest ./tests"
# dockerunix = "docker run --rm --volume \"$(pwd)\":/app --volume /app/.venv --publish 8000:8000 -it backend-api \"$@\""
# dockerwin = "docker run --rm --volume %cd%:/app --volume /app/.venv --publish 8000:8000 -it backend-api"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]
